Split TypeVisitable from TypeFoldable
